+++ Front-End/src/App.jsx
@@ -1,10 +1,16 @@
-import './App.css'
-import { Navigation } from './routes/Navigation'
+import './App.css';
+import { Navigation } from './routes/Navigation';
+import { UserProvider } from "../../Front-End/src/components/user/form 2/UserContext"
 
 const App = () => {
   return (
-    <Navigation />
-  )
-}
+    <UserProvider> 
+      <Navigation /> 
+    
+    </UserProvider>
+  );
+};
 
-export default App
+export default App;
+
+
